chore: Rename ActiveLinks to ActiveLink
